# src/Core.Models/WebAPI.Data.Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebAPI.Data",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/Display/{id}/{displayStatus}": {
      "get": {
        "tags": [
          "Display"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "displayStatus",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DisplayStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/v1/Display/{id}": {
      "delete": {
        "tags": [
          "Display"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Display"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Display"
                }
              }
            }
          }
        }
      }
    },
    "/v1/Display": {
      "delete": {
        "tags": [
          "Display"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Display"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Display"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Display"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Display"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Display"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Display"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Display"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Display"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Display"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Display"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Display"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/v1/Template/{id}": {
      "delete": {
        "tags": [
          "Template"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Template"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Template"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Display": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "ipStr": {
            "type": "string",
            "nullable": true
          },
          "macStr": {
            "type": "string",
            "nullable": true
          },
          "publicKey": {
            "type": "string",
            "nullable": true
          },
          "dummy": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/DisplayStatus"
          },
          "templateId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "groupId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "filterId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "roomId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DisplayStatus": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "Template": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "html": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateHtmlJson"
            },
            "nullable": true
          },
          "css": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateCssJson"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TemplateCssJson": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TemplateHtmlJson": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}